requirements.txt
fastapi
uvicorn[standard]
python-multipart
selenium
webdriver-manager
beautifulsoup4
pdfplumber
openai
pandas
openpyxl
PyMuPDF
pytesseract
Pillow
easyocr
pdf2image
fpdf
webdriver-manager
tiktoken
gunicorn